Move `Job` to `models` directory
Job is a SQLAlchemy model, so it makes the most sense to have it live
alongside our other db models in the `models` directory.
