+++ airflow/cli/commands/standalone_command.py
@@ -31,9 +31,9 @@ from airflow.configuration import AIRFLOW_HOME, conf, make_group_other_inaccessi
 from airflow.executors import executor_constants
 from airflow.executors.executor_loader import ExecutorLoader
 from airflow.jobs.base_job_runner import BaseJobRunner
-from airflow.jobs.job import most_recent_job
 from airflow.jobs.scheduler_job_runner import SchedulerJobRunner
 from airflow.jobs.triggerer_job_runner import TriggererJobRunner
+from airflow.models.job import most_recent_job
 from airflow.utils import db
