+++ airflow/jobs/local_task_job_runner.py
@@ -25,7 +25,7 @@ from sqlalchemy.orm import Session
 from airflow.configuration import conf
 from airflow.exceptions import AirflowException
 from airflow.jobs.base_job_runner import BaseJobRunner
-from airflow.jobs.job import Job, perform_heartbeat
+from airflow.models.job import Job, perform_heartbeat
 from airflow.models.taskinstance import TaskInstance, TaskReturnCode
 from airflow.serialization.pydantic.job import JobPydantic
 from airflow.stats import Stats
