+++ airflow/models/__init__.py
@@ -59,9 +59,9 @@ def import_all_models():
     for name in __lazy_imports:
         __getattr__(name)
 
-    import airflow.jobs.job
     import airflow.models.dagwarning
     import airflow.models.dataset
+    import airflow.models.job
     import airflow.models.serialized_dag
     import airflow.models.tasklog
     import airflow.www.fab_security.sqla.models
