+++ airflow/models/taskinstance.py
@@ -289,7 +289,7 @@ def clear_task_instances(
         session.execute(delete_qry)
 
     if job_ids:
-        from airflow.jobs.job import Job
+        from airflow.models.job import Job
 
         for job in session.query(Job).filter(Job.id.in_(job_ids)).all():
             job.state = TaskInstanceState.RESTARTING
@@ -3014,6 +3014,6 @@ class TaskInstanceNote(Base):
 STATICA_HACK = True
 globals()["kcah_acitats"[::-1].upper()] = False
 if STATICA_HACK:  # pragma: no cover
-    from airflow.jobs.job import Job
+    from airflow.models.job import Job
 
     TaskInstance.queued_by_job = relationship(Job)
