+++ docs/apache-airflow/administration-and-deployment/logging-monitoring/check-health.rst
@@ -99,7 +99,7 @@ using `http.server.BaseHTTPRequestHandler <https://docs.python.org/3/library/htt
 CLI Check for Scheduler
 -----------------------
 
-Scheduler creates an entry in the table :class:`airflow.jobs.job.Job` with information about the host and
+Scheduler creates an entry in the table :class:`airflow.models.job.Job` with information about the host and
 timestamp (heartbeat) at startup, and then updates it regularly. You can use this to check if the scheduler is
 working correctly. To do this, you can use the ``airflow jobs checks`` command. On failure, the command will exit
 with a non-zero error code.
