+++ tests/jobs/test_local_task_job.py
@@ -36,10 +36,10 @@ import pytest
 from airflow import DAG, settings
 from airflow.exceptions import AirflowException
 from airflow.executors.sequential_executor import SequentialExecutor
-from airflow.jobs.job import Job, run_job
 from airflow.jobs.local_task_job_runner import SIGSEGV_MESSAGE, LocalTaskJobRunner
 from airflow.jobs.scheduler_job_runner import SchedulerJobRunner
 from airflow.models.dagbag import DagBag
+from airflow.models.job import Job, run_job
 from airflow.models.serialized_dag import SerializedDagModel
 from airflow.models.taskinstance import TaskInstance
 from airflow.operators.empty import EmptyOperator
@@ -91,7 +91,7 @@ class TestLocalTaskJob:
     @pytest.fixture(autouse=True)
     def set_instance_attrs(self, dagbag):
         self.dagbag = dagbag
-        with patch("airflow.jobs.job.sleep") as self.mock_base_job_sleep:
+        with patch("airflow.models.job.sleep") as self.mock_base_job_sleep:
             yield
 
     def validate_ti_states(self, dag_run, ti_state_mapping, error_message):
