+++ tests/listeners/test_listeners.py
@@ -19,8 +19,8 @@ from __future__ import annotations
 import pytest as pytest
 
 from airflow import AirflowException
-from airflow.jobs.job import Job, run_job
 from airflow.listeners.listener import get_listener_manager
+from airflow.models.job import Job, run_job
 from airflow.operators.bash import BashOperator
 from airflow.utils import timezone
 from airflow.utils.session import provide_session
